Adds .gitignore file to exclude unnecessary files
This commit includes a new .gitignore file. This file lists files and directories that are not necessary to track or ought not to be part of the repository. It includes specific excludes for Python and Django, JetBrains IDEs like PyCharm, as well as for various plugins, deployment tools and project files. This prevents unnecessary bloat in the repository, and respects the privacy of individual development environments.